refactor: utils service
