--- src/app/shared/utils/utils.ts
@@ -1,8 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-export const range = (start: number, end: number): number[] => {
-  return [...Array(end - start).keys()].map(el => el + start);
-};
-
-export const pluck = (elements: any[], field: string) => {
-  return elements.map(el => el[field]);
-};
